Feat finalização task voto
Feat: Concluído a task para contabilização dos votos das receitas e criado endpoint para retornar o usuario por email

+++ CookMaster-APP/cookmasterapi/src/main/java/br/integration/cookmasterapi/dto/ReceitaDto.java
@@ -19,7 +19,7 @@ public class ReceitaDto extends BaseDto<Receita, ReceitaDto> {
     private boolean ativo;
     private String imagem;
     private Usuario usuario;
-    private Double voto;
+    private int voto;
     private Categoria categoria;
 
     @Override
